assets/agent/requirements.txt
litellm==1.81.0,!=1.82.7,!=1.82.8
langchain==1.2.6
langchain-core==1.2.28
langchain-litellm==0.3.5
langgraph==1.0.10
a2a-sdk[all]==0.3.22
uvicorn==0.40.0
httpx==0.28.1
python-dotenv==1.2.1
click==8.3.1
python-dateutil==2.9.0
python-json-logger>=3.3.0
sap-cloud-sdk==0.6.1.dev20260410+contribs-sapphire-internal-testing.42f21d5
